# openapi.yaml
openapi: 3.0.0
info:
  title: AIQuiz API
  description: Complete API documentation for the AIQuiz educational platform with AI-powered question generation and RAG-enabled document processing
  version: 1.0.0
  contact:
    name: AIQuiz Team
servers:
  - url: /api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/manager/auth/login
  schemas:
...
    Questionnaire:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            type: string
        topic:
          type: string
        createdBy:
          type: string
...
paths:
...
  /manager/subjects/{id}/topics/{topicId}/questionnaires/{questionnaireId}:
    get:
      summary: Get questionnaire details
      tags:
        - Questionnaires
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Questionnaire details with populated questions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Questionnaire'

    put:
      summary: Update questionnaire
      tags:
        - Questionnaires
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                questions:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Questionnaire updated successfully

    delete:
      summary: Delete questionnaire
      tags:
        - Questionnaires
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
        - name: questionnaireId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Questionnaire deleted successfully
...